requirements.txt
flask
werkzeug
cloudinary
psycopg2-binary
requests
